requirements.txt
rdkit>=2023.9.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
biopython>=1.83
pillow>=10.0.0
jupyter>=1.0.0
requests>=2.31.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.65.0
reportlab>=4.0.0
streamlit>=1.28.0
py3Dmol>=2.0.0
plotly>=5.14.0
